Make types api routes check their privilege.

+++ backend/src/controllers/type_controller.js
@@ -1,8 +1,11 @@
 const router = require('express').Router()
 
 const typeService = require('../services/type_service.js')
+const { checkPrivilege } = require('../services/auth.js')
+const globalMessages = require('../messages/global_messages.js')
 
 const messages = {
+  ...globalMessages,
   create: {
     success: {
       eng: '"Tyyppi luotu onnistuneesti." englanniksi.',
@@ -20,7 +23,12 @@ const messages = {
 }
 
 router.post('/create', async (req, res) => {
-  // TODO validate
+  if (!checkPrivilege(req, ['logged_in', 'teacher_on_course'])) {
+    res.status(403).json({
+      error: messages.privilege.failure[req.lang]
+    })
+    return
+  }
   const created = await typeService.create(req.body, req.lang)
   res.status(200).json({
     message: messages.create.success[req.lang],
@@ -29,7 +37,12 @@ router.post('/create', async (req, res) => {
 })
 
 router.delete('/:id', async (req, res) => {
-  // TODO validate
+  if (!checkPrivilege(req, ['logged_in', 'teacher_on_course'])) {
+    res.status(403).json({
+      error: messages.privilege.failure[req.lang]
+    })
+    return
+  }
   const deleted = await typeService.delete(Number(req.params.id))
   res.status(200).json({
     message: messages.delete.success[req.lang],
